src/orchestrator/requirements.txt
semantic-kernel
fastapi
uvicorn
pydantic
azure-storage-blob
psycopg2-binary
# mcp